requirements.txt
Django==1.11.15
django-extensions==2.1.2
django-extra-views==0.11.0
Pillow==5.3.0
psycopg2-binary==2.7.5
six==1.11.0
